# docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Lanchonete
  version: 1.0.0
  description: >
    API REST para gerenciamento de pedidos,
    produtos, estoque, usuários, pagamentos
    e unidades da lanchonete.
servers:
  - url: https://lanchonete-api.onrender.com
    description: Servidor de produção

  - url: http://127.0.0.1:5000
    description: Servidor local
...
paths:
...
  /admin/register/produto:
    post:
      tags:
      - Registros
      - Produtos
      summary: Realizar cadastro de produtos
      description: Cadastra produtos.
      operationId: cadastrarProdutos
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - nome
              - preco
              - categoria
              - descricao
              properties:
                nome:
                  type: string
                  description: Nome do produto.
                  example: "Pão de queijo"
                preco:
                  type: number
                  format: double
                  description: Preço do produto, válido para todas as unidades.
                  example: 15.5
                categoria:
                  type: string
                  description: Categoria em que o produto se encaixa.
                  example: "bebidas"
                descricao:
                  type: string
                  description: Breve descrição sobre o produto ofertado.
                  example: "Bebida feita à base de guaraná."
      responses:
        "201":
          description: Cadastro realizado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "O produto foi cadastrado com sucesso."
                data:
                  id: 2
                  nome: Café pequeno
                  categoria: bebidas
                  preco: "3.20"
                  descricao: Café preto coado, xícara de 200ml
        "400":
          description: Requisição inválida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "JSON_INVALIDO"
                message: "O corpo da requisição deve ser um JSON válido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/register/produto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "409":
          description: Conflito com o estado atual do recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_JA_CADASTRADO"
                message: "O produto informado já se encontra cadastrado."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/register/produto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "VALOR_INVALIDO"
                message: "O preço é obrigatório."
                details: [{field: "preco", issue: "Informe o preço do produto."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/register/produto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/register/produto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        '401':
          description: Token JWT ausente, expirado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Usuário autenticado sem permissão para acessar este recurso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /admin/produto/{produtoId}/valor:
    patch:
      tags:
      - Produtos
      summary: Alterar o valor do produto.
      description: >
        Permite atualizar o valor do produto desejado. Somente a gestão pode realizar esta ação.  
      operationId: atualizarValor
      security:
      - bearerAuth: []
      parameters:
      - name: produtoId
        in: path
        required: true
        description: Produto que terá seu valor mudado.
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - novoValor
              properties:
                novoValor:
                  type: number
                  format: double
                  description: Novo valor do produto.
                  example: 13.5
      responses:
        "200":
          description: Produto atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "O preço foi atualizado em todas as unidades."
                data:
                  produtoId: 2
                  produto: "pao de queijo"
                  novoValor: "13.50"
                  estoquesAtualizados: 3
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/produto/2/valor"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para realizar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/produto/2/valor"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este produto."
                details: [{field: "produtoId", issue: "O produto com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/produto/2/valor"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "NOVO_VALOR_INVALIDO"
                message: "O valor inserido é inválido."
                details: [{field: "novoValor", issue: "O novo valor deve ser maior ou igual a 0,00."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/produto/2/valor"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/produto/2/valor"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /funcionarios/precos/alteracoes: ##########
    get:
      tags:
      - Produtos
      summary: Listar alterações de preço
      description: >
        Consulta de alterações de preço no sistema com possibilidade de utilizar filtros de pesquisa, escolher a ordenação
        e a paginação.
      security:
      - bearerAuth: []
      parameters:
      - name: produtoId
        in: query
        required: false
        description: Filtra pelo ID do produto.
        schema:
          type: integer
          minimum: 1
      - name: nomeProduto
        in: query
        required: false
        description: Filtra pelo nome do produto.
        schema:
          type: string
      - name: categoria
        in: query
        required: false
        description: Filtra pela categoria do produto.
        schema:
          type: string
      - name: dataInicio
        in: query
        required: false
        description: Filtra a partir da data informada.
        schema:
          type: string
          format: date
        example: "YYYY-MMM-DD"
      - name: dataFim
        in: query
        required: false
        description: Filtra produtos até a data final informada.
        schema:
          type: string
          format: date
        example: "2026-08-01"
      - name: novoValorMin
        in: query
        required: false
        description: Filtra alterações com o menor valor informado.
        schema:
          type: number
          format: double
          minimum: 0
      - name: novoValorMax
        in: query
        required: false
        description: Filtra alterações com o maior valor informado.
        schema:
          type: number
          format: double
          minimum: 0
      - name: ordenacao
        in: query
        required: false
        description: Define a ordenação dos resultados.
        schema:
          type: string
          default: promocaoId_asc
          enum:
          - promocaoId_asc
          - promocaoId_desc
          - nomePromo_asc
          - nomePromo_desc
          - dataPromo_asc
          - dataPromo_desc
      - name: page
        in: query
        required: false
        description: Número da página.
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: limit
        in: query
        required: false
        description: Quantidade máxima de registros por página.
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        "200":
          description: Lista de alterações consultada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricoPrecosResponse'
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/funcionarios/precos/alteracoes"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para acessar este conteúdo."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/funcionarios/precos/alteracoes"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Um ou mais filtros informados são inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "FILTRO_INVALIDO"
                message: "O filtro informado é inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/funcionarios/precos/alteracoes"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/funcionarios/precos/alteracoes"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
      operationId: consultaAlteracao
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false
    ApiError:
      type: object
      properties:
        error:
          type: string
          example: "USUARIO_NAO_ENCONTRADO"
        message:
          type: string
          example: "O usuário informado não foi encontrado."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "usuario"
              issue:
                type: string
                example: "O usuário informado não existe."
        timestamp:
          type: string
          format: date-time
          example: "2026-08-07T08:30:00-03:00"
        path:
          type: string
          example: "/login"
        requestId:
          type: string
          format: uuid
          example: "baaf6e1b-5217-41af-9877-2298562e65e3"
    HistoricoPreco:
      type: object
      properties:
        historicoId:
          type: integer
          example: 5
        produtoId:
          type: integer
          example: 6
        usuarioId:
          type: integer
          example: 3
        precoAnterior:
          type: string
          example: "13.99"
        precoNovo:
          type: string
          example: '14.55'
        example: "14.55"
        dataAlteracao:
          type: string
          format: date-time
          example: "2026-08-01T10:30:00"
    RespostaSucesso:
      type: object
      properties:
        message:
          type: string
          example: Operação realizada com sucesso.
        data: {}
...
    HistoricoPrecosResponse:
      type: object
      properties:
        message:
          type: string
          example: Lista de alterações consultada com sucesso.
        data:
          type: array
          items:
            $ref: '#/components/schemas/HistoricoPreco'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...